pyproject.toml
[tool.poetry]
name = "tm-density-calculator-v1"
version = "0.1.0"
description = ""
authors = ["smbrine <smbrine@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
shapely = "^2.0.4"
geopandas = "^0.14.4"
overpass = "^0.7"
geopy = "^2.4.1"
pyproj = "^3.6.1"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.3.1"
grpcio = "^1.64.1"
grpcio-tools = "^1.64.1"
grpcio-reflection = "^1.64.1"
osm2geojson = "^0.2.5"
area = {git = "https://github.com/smbrine/area.git", rev = "master", tag='v1.1.3'}


[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
black = "^24.4.2"
pylint = "^3.2.2"
flake8 = "^7.0.0"
ruff = "^0.4.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
